Move kubernetes functions out into a util/kubeclient package
This is to make the code more testable.
Currently there is too much to mock out in cluster, this makes the tests very
thick and difficult to manage, by placing the functions that call kubernetes
behind an interface we can mock them out in the cluster tests making them simpler.