Follow-up to https://github.com/openshift/openshift-azure/pull/818
Avoid storing the Blob object in the simpleUpgrader, evidently this was a
mistake.  On a Blob.Get() the blob MD5 is latched into the Blob object.  If
Blob.CreateBlockBlobFromReader() is called, the latched MD5 is not cleared,
resulting in an error on the subsequent Get().  Store the Container in the
simpleUpgrader to force the user to get the Blob themselves.