Fix fakerp.FindDirectory to find folders at system root
I discovered this while working out how to run our e2e tests from a
container. walkParents was checking if the path we are analysing is the
system root (/) and exits if it is. The problem with this is that the
sought-after directory could be at root directory e.g. in a container at
/_data

This fix reverses the order of the stop condition in walkParents and allows
an e2e binary in a container (at /e2e) to find the _data or secrets
directories at /_data or /secrets in the container.

/cc @kargakis @mjudeikis