Adding an dedicated drive for docker data
We're currently using the temporary instance drive for the docker data
on each of the nodes. This is a variable drive (the bigger the instance,
the bigger the drive) and since we want to support several instance
types ranging from fairly small to big-ish, it's hard to guarantee the
same space will be available. OCD is reserving 200G space for their
docker data, I am going to 256 to take advantage of dedicated IOPs in
azure disks.