Rename azure client files to match their azure packages and interfaces
This also tries to standardize on importing azure packages with "az" when there
is a conflict with one our packages. This will happen a lot now as we are
naming the packages the same as azure's.

The mock generation is moved to each sub package of azureclient

The azure client fakes are also changed to the new standard

This required a change to the way AzureCloud was passed into the fakeClients
to break an import loop.