Moving geneva logging stack
Previously, was running the mdsd/td-agent docker image as a deployment
in the openshift-azure-logging namespace. As a requirement for azsecpack
we need to run geneva on the host themselves, this means that the
openshift-azure-logging namespace is going away and the geneva logging
stack is installed in the vmimage build process, and the config parsing
logic moves to the startup container.

This is a squashed commit, below is the rest of the commits squashed:

Add mdsd/td-agent/azsecpack to the vmimage

Should work so the services don't come up until they are properly
configured

Add GenevaLogging certs/config to worker nodes

They do require it just like masters do

Populate the geneva logging stack config

Need to lay down the certificates and environment variables for geneva
and td-agent to startup properly

Add the name of the cluster to the workers

so it's present for geneva config

Attempt to monitor only the certs and cfg changes

Switch to a derive pattern

No need to have these repeated all over the place

Update with better docs

The logic moved to derive.go

No longer needed here

Removing unnecessary debug statement

Using a common/ directory

let's avoid repeating config files in > 1 place

Replacing td-agent with fluentd

This means we don't have to build the systemd input gem, still
troubleshooting this.

Fix the fluentd config

* Removed options that matched the default values
* Changed the heartbeat: the previous incarnation of the td-agent wasn't
  talking directly to mdsd (which now understands native fluent
  protocol) but the mdsd server does not listen on UDP port, it was
  marking it as gone. The buffer settings are taken from the previous
  mdsd type config, as both the forward and mdsd types are buffered
  plugins

Forgot to properly identify systemd units

regenerated bindata

Using a slice to guarantee the file map is looped in order

unmask first, then enable