Improves make vmimage-validate locally
Currently scap report and yum log files are not being cleaned up between `make vmimage-validate` runs.
It's not an issue in CI, because we run jobs in clean environments,
but can cause confusion when run locally because the validation will fail
even if the image is valid. It happens because the command checks if files contain data or not.

Example scenario:

* You run validation locally and it fails due to pending security updates
* You fix the issue/update vm image
* You run validation locally again: it fails becasue `yum_updateinfo_list_security` log contains data from previous run