Disable DenyEscalatingExec
The alternative to this quick fix would be to:

- remove denyescalatingexec from master config
- update admission controller registration such that it gets called to
  validate CONNECTs to pods
- add wiring in the admission controller such that those validation
  calls forward to the relevant function in the origin
  denyescalatingexec codebase
- insert a check up front that detects the binary build case and
  override allows it

This is significantly more work, was investigated by Pris for
feasibility, and me/Jim/Pris all concluded that the committed fix is
simpler and will be easier to maintain.

We do not believe this to be a highly risky change. The
ValidatingAdmissionWebhook only permits the creation of privileged
containers in very limited circumstances; this change allows a
customer-admin user to exec into one of those privileged, whitelisted
pods.

Co-authored-by: Pris Nasrat <pnasrat@redhat.com>